* Added Board to handle the view of the game (creation of the board, grid, scores, overlay,...)
* GameManager just takes care of the game logic.
